fix tidy

+++ src/include/migraphx/op/pack_int4.hpp
@@ -91,7 +91,7 @@ struct pack_int4
                     in_data_multi_idx[axis] *= 2;
                     auto input_val = input[in_data_multi_idx];
                     // mask first 4 bits, keep it little endian.
-                    output[i] = 0x0F & input_val;
+                    output[i] = static_cast<uint8_t>(0x0F) & input_val;
                     in_data_multi_idx[axis] += 1;
                     input_val = input[in_data_multi_idx];
                     output[i] = (input_val << 4) | (output[i]);
